+++ src/model/graph/Edge.java
@@ -5,14 +5,13 @@ public class Edge {
     private Node a, b;
     private String type;
     private boolean flooded;
-    private double value;
+    private int value;
 
     public Edge(Node a, Node b, String type) {
         this.a = a;
         this.b = b;
         this.type = type;
         this.flooded = false;
-//        this.value = Math.sqrt(Math.pow((b.getX() - a.getX()), 2) + Math.pow((b.getY() - a.getY()), 2));
         this.updateValue();
     }
 
@@ -32,19 +31,17 @@ public class Edge {
         return flooded;
     }
 
-    public double getValue() {
+    public int getValue() {
         return value;
     }
 
     public void setA(Node a) {
         this.a = a;
-//        this.value = Math.sqrt(Math.pow((a.getX() - a.getY()), 2) + Math.pow((b.getX() - b.getY()), 2));
         this.updateValue();
     }
 
     public void setB(Node b) {
         this.b = b;
-//        this.value = Math.sqrt(Math.pow((a.getX() - a.getY()), 2) + Math.pow((b.getX() - b.getY()), 2));
         this.updateValue();
     }
 
@@ -60,7 +57,7 @@ public class Edge {
      * Met a jour la valeur de l'edge
      */
     private void updateValue(){
-        this.value = Math.sqrt(Math.pow((b.getX() - a.getX()), 2) + Math.pow((b.getY() - a.getY()), 2));
+        this.value = (int) Math.sqrt(Math.pow((b.getX() - a.getX()), 2) + Math.pow((b.getY() - a.getY()), 2));
     }
 
     @Override
